% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{gen_rand_order}
\alias{gen_rand_order}
\title{Function to generate a stepped-wedge cRCT randomisation allocation}
\usage{
gen_rand_order(nJ, nT)
}
\arguments{
\item{nJ}{Number of clusters}

\item{nT}{Number of time points}
}
\value{
A data frame with columns cl and t indicating the time
}
\description{
Function to generate a stepped-wedge cRCT randomisation allocation. Assumes a baseline
and endline period in which no clusters and all clusters have the intervention,
respectively.
}
