% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststats.R
\name{get_G}
\alias{get_G}
\title{Get inverse partial derivative of link function}
\usage{
get_G(x, family)
}
\arguments{
\item{x}{vector of predicted model values on the scale of the linear predictor}

\item{family}{a family object, see \link[stats]{family}. Only identity, log, logit, and probit link functions currently supported.}
}
\value{
A vector of values of the inverse of the partial derivative of the link function with respect to the linear predictor
}
\description{
Returns the inverse partial derivative of the link function with respect to the linear predictor
}
\examples{
get_G(seq(-1,1,length.out=10),family=binomial())
}
