% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{perm_dist}
\alias{perm_dist}
\title{Extracts the test statistics}
\usage{
perm_dist(out, positions)
}
\arguments{
\item{out}{Array output by the \code{permute} function}

\item{positions}{Vector indicating which rows of out to use}
}
\description{
Extracts the test statistics from the output of the \code{permute} function.
Returns the largest value from a specified subset of rows, each row is the test
statistic from a different null hypothesis.
}
