% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simpleLM}
\alias{simpleLM}
\title{A very basic linear model solver}
\usage{
simpleLM(y_, X_)
}
\arguments{
\item{y_}{A vector of outcome values}

\item{X_}{The design matrix of fixed effects}
}
\value{
A list with the parameter values and fitted values
}
\description{
Returns the OLS paramter estimates and fitted values. Used internally for quick
fitting of null models.
}
