% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permute}
\alias{permute}
\title{Wrapper function to replicate permutations}
\usage{
permute(
  fitlist,
  data,
  n_permute = 100,
  null_pars = rep(0, length(fitlist)),
  cl_var = "cl",
  tr_var = "treat",
  rand_func = NULL,
  inv_sigma = NULL
)
}
\arguments{
\item{fitlist}{A list of glm model objects fitted under the null hypotheses}

\item{data}{A data frame containing the data used to fit the models in fitlist}

\item{n_permute}{Number of permutations to conduct}

\item{null_pars}{A vector of the same length as fitlist specifying the value(s) of the
treatment effect parameter(s) under the null hypotheses}

\item{cl_var}{String specifying the name of the column identifying the treatment allocation}

\item{rand_func}{The name of a function that re-randomises the clusters. The function should
produce a data frame that identifies the clusters in the treatment group under the new
randomisation scheme. The data frame can either have a single column with name cl_var or
two columns of cl_var and t identifying the cluster ID and time period a cluster joins
the treatment group.  If NULL then clusters are randomised in a 1:1 ratio to treatment and control}

\item{inv_sigma}{optional, inverse of the estimated covariance matrix of the observations. If provided then the weighted q-score statistic
is calculated}
}
\value{
An array of dimension length(fitlist)*n_permute containing the test statistics
for each model and each iteration
}
\description{
Calls lme_permute2 to replicate the permutations
}
\examples{
out <- twoarm_sim()
data <- out[[1]]
  fit1 <- lme4::glmer(y1 ~ treat + (1|cl) ,
data=data,
family="poisson")

fit2 <- lme4::glmer(y2 ~ treat + (1|cl),
                    data=data,
                    family="poisson")
fitlist <- list(fit1,fit2)
nullfitlist <- list()
for(i in 1:length(fitlist)){
  nullfitlist[[i]] <- est_null_model(fitlist[[i]],
                                     data,
                                     tr_var = "treat",
                                     null_par = 0)
}
out <- permute(nullfitlist,
               data=data,
               n_permute = 10,
               cl_var = "cl")
}
